pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "openconnect-saml"
version = "0.8.0"
description = "OpenConnect wrapper with Azure AD (SAML) SSO support for Cisco SSL-VPNs"
readme = "README.md"
license = "GPL-3.0-or-later"
requires-python = ">=3.10"
authors = [
    { name = "Matthias Schabhüttl" },
]
keywords = ["openconnect", "vpn", "saml", "azure-ad", "sso", "cisco"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: X11 Applications :: Qt",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: POSIX :: Linux",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: System :: Networking",
]
dependencies = [
    "attrs",
    "colorama",
    "keyring",
    "lxml",
    "prompt-toolkit",
    "pyotp",
    "pysocks",
    "pyxdg",
    "requests>=2.33.0",
    "structlog",
    "toml",
]

[project.optional-dependencies]
gui = [
    "pyqt6",
    "pyqt6-webengine",
]
chrome = [
    "playwright",
]
fido2 = [
    "fido2>=1.0",
]
tui = [
    "rich>=13.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=1.0",
    "pytest-cov>=5.0",
    "pytest-httpserver>=1.0",
    "ruff>=0.4",
]

[project.scripts]
openconnect-saml = "openconnect_saml.cli:main"

[project.urls]
Homepage = "https://github.com/mschabhuettl/openconnect-saml"
Repository = "https://github.com/mschabhuettl/openconnect-saml"
"Bug Tracker" = "https://github.com/mschabhuettl/openconnect-saml/issues"
Changelog = "https://github.com/mschabhuettl/openconnect-saml/releases"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["openconnect_saml"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.bandit]
exclude_dirs = ["tests"]
skips = ["B101"]  # assert used in tests is fine

[tool.coverage.run]
source = ["openconnect_saml"]
omit = ["openconnect_saml/browser/webengine_process.py"]  # Qt GUI needs display

[tool.coverage.report]
show_missing = true
fail_under = 50
